Moved Google maps to contact page
Moved Google maps and fixed other small format settings
