+++ src/main/java/ie/cit/architect/protracker/gui/ClientBilling.java
@@ -35,7 +35,7 @@ public class ClientBilling
         Scene scene = new Scene(pane, Consts.APP_WIDTH, Consts.APP_HEIGHT);
         scene.getStylesheets().add("/stylesheet.css");
         stage.setScene(scene);
-        stage.setTitle(Consts.APPLICATION_TITLE + "Billing");
+        stage.setTitle(Consts.APPLICATION_TITLE + " Billing");
         stage.show();
     }
