+++ src/main/java/ie/cit/architect/protracker/gui/ClientViewProjectTimeline.java
@@ -10,11 +10,7 @@ import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.*;
-import javafx.scene.web.WebEngine;
-import javafx.scene.web.WebView;
 import javafx.stage.Stage;
-
-import java.net.URL;
 
 /**
  * Created by Adam on 04/03/2017.
@@ -33,7 +29,7 @@ public class ClientViewProjectTimeline
         BorderPane pane = new BorderPane();
         pane.setTop(homeButtonContainer());
 
-        pane.setCenter(new Browser());
+        pane.setCenter(createProjectTimeline());
         pane.setPrefSize(300, 300);
 
         pane.setBottom(navButtonContainer());
@@ -41,11 +37,9 @@ public class ClientViewProjectTimeline
         Scene scene = new Scene(pane, Consts.APP_WIDTH, Consts.APP_HEIGHT);
         scene.getStylesheets().add("/stylesheet.css");
         stage.setScene(scene);
-        stage.setTitle(Consts.APPLICATION_TITLE + "Project Timeline");
+        stage.setTitle(Consts.APPLICATION_TITLE + " Project Timeline");
         stage.show();
     }
-
-
 
     private Pane createProjectTimeline()
     {
@@ -142,26 +136,6 @@ public class ClientViewProjectTimeline
         vb1.setAlignment(Pos.TOP_RIGHT);
 
         return vb1;
-    }
-
-    class Browser extends Region {
-
-
-        final WebView browser = new WebView();
-        final WebEngine webEngine = browser.getEngine();
-
-        public Browser() {
-
-            getStyleClass().add("browser");
-
-            final URL urlGoogleMaps = getClass().getResource("/GoogleMapsV3.html");
-            webEngine.load(urlGoogleMaps.toExternalForm());
-            webEngine.setJavaScriptEnabled(true);
-
-
-            getChildren().add(browser);
-
-        }
     }
 }
