+++ src/main/resources/GoogleMapsV3.html
@@ -3,9 +3,9 @@
 <head>
     <meta name="viewport" content="initial-scale=1.0, user-scalable=no" />
     <style type="text/css">
-        html { height: 100% }
-        body { height: 100%; margin: 0; padding: 0 }
-        #map_canvas { height: 100% }
+        html { height: 50% }
+        body { height: 50%; margin: 0; padding: 0 }
+        #map_canvas { height: 50% }
     </style>
     <script type="text/javascript"
             src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDZj4CsOZWwx0UZeLHNwO5iOQsfchBNAlM&sensor=false">
